# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mad Factory - Factory Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /supplier-payments:
    get:
      operationId: listSupplierPayments
      tags: [payments]
      parameters:
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of supplier payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierPayment"
    post:
      operationId: createSupplierPayment
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierPaymentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPayment"
  /supplier-payments/{id}:
    get:
      operationId: getSupplierPayment
      tags: [payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPayment"
    delete:
      operationId: deleteSupplierPayment
      tags: [payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Exchange Rates ─────────────────────────────────────────────────────────
...
  /shareholders:
    get:
      operationId: listShareholders
      tags: [shareholders]
      responses:
        "200":
          description: List of shareholders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shareholder"
    post:
      operationId: createShareholder
      tags: [shareholders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShareholderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shareholder"
  /shareholders/{id}:
    get:
      operationId: getShareholder
      tags: [shareholders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shareholder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shareholder"
    patch:
      operationId: updateShareholder
      tags: [shareholders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShareholderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shareholder"
    delete:
      operationId: deleteShareholder
      tags: [shareholders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /shareholder-transactions:
    get:
      operationId: listShareholderTransactions
      tags: [shareholders]
      parameters:
        - name: shareholderId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of shareholder transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShareholderTransaction"
    post:
      operationId: createShareholderTransaction
      tags: [shareholders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShareholderTransactionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareholderTransaction"
  /shareholder-transactions/{id}:
    delete:
      operationId: deleteShareholderTransaction
      tags: [shareholders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Cashbox / Financial Dashboard ─────────────────────────────────────────
...
components:
  schemas:
...
    SupplierPayment:
      type: object
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        amount:
          type: number
        currency:
          type: string
        exchangeRateId:
          type: ["integer", "null"]
        amountIqd:
          type: number
        paymentDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - supplierId
        - supplierName
        - amount
        - currency
        - amountIqd
        - paymentDate
        - createdAt
    CreateSupplierPaymentBody:
      type: object
      properties:
        supplierId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        exchangeRateId:
          type: ["integer", "null"]
        paymentDate:
          type: string
        voucherType:
          type: string
          enum: [cash, internal]
        notes:
          type: ["string", "null"]
      required:
        - supplierId
        - amount
        - currency
        - paymentDate

    # ── Exchange Rate ─────────────────────────────────────────────────────────
...
    Shareholder:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sharePercentage:
          type: number
        shareAmount:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        totalWithdrawn:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - sharePercentage
        - isActive
        - totalWithdrawn
        - createdAt
        - updatedAt
    CreateShareholderBody:
      type: object
      properties:
        name:
          type: string
        sharePercentage:
          type: number
        shareCount:
          type: ["number", "null"]
        shareAmount:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - sharePercentage
    UpdateShareholderBody:
      type: object
      properties:
        name:
          type: string
        sharePercentage:
          type: number
        shareAmount:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
    ShareholderTransaction:
      type: object
      properties:
        id:
          type: integer
        shareholderId:
          type: integer
        shareholderName:
          type: string
        type:
          type: string
        amount:
          type: number
        transactionDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - shareholderId
        - shareholderName
        - type
        - amount
        - transactionDate
        - createdAt
    CreateShareholderTransactionBody:
      type: object
      properties:
        shareholderId:
          type: integer
        type:
          type: string
        amount:
          type: number
        transactionDate:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - shareholderId
        - type
        - amount
        - transactionDate

    # ── Account Statement (shared) ────────────────────────────────────────────
...